backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
scikit-learn
